% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_utils.R
\name{xg_define_disease}
\alias{xg_define_disease}
\title{Define Disease by ICD-10}
\usage{
xg_define_disease(df, icd_codes, strict = FALSE)
}
\arguments{
\item{df}{Loaded data frame (must contain f.41270 series columns).}

\item{icd_codes}{Character vector, e.g., c("I20", "I25").}

\item{strict}{Logical. TRUE = exact match, FALSE = prefix match (default).}
}
\value{
Returns a vector consisting of 0s and 1s, where 1 indicates the presence of the disease.
}
\description{
Takes ICD prefixes as input and automatically identifies diagnosed patients from the main and secondary diagnosis columns (f.41270).
}
\author{
Xu Jun <xujun05@pku.edu.cn>
}
