% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_utils.R
\name{xg_get_cols}
\alias{xg_get_cols}
\title{Get Complete Column Names for a Specific Field ID}
\usage{
xg_get_cols(field_id, data_cols)
}
\arguments{
\item{field_id}{Character or numeric.}

\item{data_cols}{A vector containing all column names of the dataset.}
}
\value{
A vector of matched column names.
}
\description{
Takes a Field ID (e.g., "21001") as input and returns the actual column names from the dataset (e.g., "f.21001.0.0").
}
