% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_rename.R
\name{xg_rename}
\alias{xg_rename}
\title{Rename columns based on UKB Dictionary (Field ID or Name)}
\usage{
xg_rename(data, dict = ukb_dict_mini)
}
\arguments{
\item{data}{A data.frame or data.table with raw column names.}

\item{dict}{The dictionary data frame. MUST contain 'Description'. 
Should contain 'FieldID' and/or 'Name'.}
}
\value{
The data with renamed columns.
}
\description{
Rename columns in a dataset using a UKB dictionary. 

For Olink data, it extracts the short protein name from the Description 
(the part before the first semicolon).
}
\author{
Jun Xu <xujun05@pku.edu.cn>
}
