% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_utils.R
\name{xg_find}
\alias{xg_find}
\title{搜索变量字典 (Search Variable Dictionary)}
\usage{
xg_find(keywords, view = TRUE)
}
\arguments{
\item{keywords}{一个或多个字符串关键词 (如 c("sex", "age"))。支持中文或英文（取决于字典内容）。}

\item{view}{逻辑值，是否在 View 窗口中展示结果。默认为 TRUE。}
}
\value{
返回包含 FieldID, Description 等信息的 data.frame
}
\description{
根据关键词模糊搜索 UK Biobank 变量字典，无需打开网页版 Showcase。
}
\examples{
\dontrun{
xg_find("diabetes")
xg_find(c("血压", "blood pressure"))
}
}
