% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_utils.R
\name{xg_get_cols}
\alias{xg_get_cols}
\title{获取指定 Field ID 的完整列名}
\usage{
xg_get_cols(field_id, data_cols)
}
\arguments{
\item{field_id}{字符串或数字}

\item{data_cols}{数据集的所有列名向量}
}
\value{
匹配到的列名向量
}
\description{
输入 Field ID (如 "21001")，返回数据集中实际的列名 (如 "f.21001.0.0")
}
