% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg_rename.R
\name{xg_rename}
\alias{xg_rename}
\title{Rename columns from UKB Field IDs to Descriptions}
\usage{
xg_rename(data, dict = ukb_dict_mini)
}
\arguments{
\item{data}{A data.frame or data.table with raw column names (e.g., p21001_i0).}

\item{dict}{The dictionary data frame (e.g., ukb_dict_mini).}
}
\value{
The data with renamed columns.

The data with renamed columns.
}
\description{
Rename columns from UKB Field IDs to Descriptions
}
